popravljeni neki eclipse warningsi
git-svn-id: http://morgoth.zemris.fer.hr/svn/vhdllab/vhdllab2/trunk@1238 cbae1e92-611c-0410-ab7b-b19ac7dee622

+++ vhdllab-common/src/main/java/hr/fer/zemris/vhdllab/applets/editor/schema2/misc/Time.java
@@ -126,7 +126,7 @@ public final class Time {
 
 	@Override
 	public final String toString() {
-		return (double) this.timeInterval + " " + this.timeMetric.toString();
+		return this.timeInterval + " " + this.timeMetric.toString();
 	}
 
 }
